refactor(experiment.go): extract build and list of experiments
This change extracts from experiment.go the ExperimentBuilder struct
as well as the list of all the experiments. It makes working with
experiment.go more manageable by moving separate components into two
distinct separate files. This change is yak shaving while I am
investigating https://github.com/ooni/explorer/issues/495.
