+++ experiment.go
@@ -5,28 +5,11 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"net/http"
 	"os"
-	"reflect"
 	"strconv"
 	"time"
 
-	"github.com/iancoleman/strcase"
-	"github.com/ooni/probe-engine/experiment/dash"
-	"github.com/ooni/probe-engine/experiment/example"
-	"github.com/ooni/probe-engine/experiment/fbmessenger"
-	"github.com/ooni/probe-engine/experiment/hhfm"
-	"github.com/ooni/probe-engine/experiment/hirl"
-	"github.com/ooni/probe-engine/experiment/ndt7"
-	"github.com/ooni/probe-engine/experiment/psiphon"
-	"github.com/ooni/probe-engine/experiment/sniblocking"
-	"github.com/ooni/probe-engine/experiment/stunreachability"
-	"github.com/ooni/probe-engine/experiment/telegram"
-	"github.com/ooni/probe-engine/experiment/tor"
-	"github.com/ooni/probe-engine/experiment/urlgetter"
-	"github.com/ooni/probe-engine/experiment/webconnectivity"
-	"github.com/ooni/probe-engine/experiment/whatsapp"
 	"github.com/ooni/probe-engine/internal/platform"
 	"github.com/ooni/probe-engine/model"
 	"github.com/ooni/probe-engine/netx/bytecounter"
@@ -40,161 +23,6 @@ const dateFormat = "2006-01-02 15:04:05"
 
 func formatTimeNowUTC() string {
 	return time.Now().UTC().Format(dateFormat)
-}
-
-// InputPolicy describes the experiment policy with respect to input. That is
-// whether it requires input, optionally accepts input, does not want input.
-type InputPolicy string
-
-const (
-	// InputRequired indicates that the experiment requires
-	// external input to run. If this input is not provided to
-	// the experiment, it will not know what to do.
-	InputRequired = InputPolicy("required")
-
-	// InputOptional indicates that the experiment handles input,
-	// if any; otherwise it fetchs input/uses a default.
-	InputOptional = InputPolicy("optional")
-
-	// InputNone indicates that the experiment does not want any
-	// input and ignores the input if provided with it.
-	InputNone = InputPolicy("none")
-)
-
-// ExperimentBuilder is an experiment builder.
-type ExperimentBuilder struct {
-	build         func(interface{}) *Experiment
-	callbacks     model.ExperimentCallbacks
-	config        interface{}
-	inputPolicy   InputPolicy
-	interruptible bool
-}
-
-// Interruptible tells you whether this is an interruptible experiment. This kind
-// of experiments (e.g. ndt7) may be interrupted mid way.
-func (b *ExperimentBuilder) Interruptible() bool {
-	return b.interruptible
-}
-
-// InputPolicy returns the experiment input policy
-func (b *ExperimentBuilder) InputPolicy() InputPolicy {
-	return b.inputPolicy
-}
-
-// OptionInfo contains info about an option
-type OptionInfo struct {
-	Doc  string
-	Type string
-}
-
-// Options returns info about all options
-func (b *ExperimentBuilder) Options() (map[string]OptionInfo, error) {
-	result := make(map[string]OptionInfo)
-	ptrinfo := reflect.ValueOf(b.config)
-	if ptrinfo.Kind() != reflect.Ptr {
-		return nil, errors.New("config is not a pointer")
-	}
-	structinfo := ptrinfo.Elem().Type()
-	if structinfo.Kind() != reflect.Struct {
-		return nil, errors.New("config is not a struct")
-	}
-	for i := 0; i < structinfo.NumField(); i++ {
-		field := structinfo.Field(i)
-		result[field.Name] = OptionInfo{
-			Doc:  field.Tag.Get("ooni"),
-			Type: field.Type.String(),
-		}
-	}
-	return result, nil
-}
-
-// SetOptionBool sets a bool option
-func (b *ExperimentBuilder) SetOptionBool(key string, value bool) error {
-	field, err := fieldbyname(b.config, key)
-	if err != nil {
-		return err
-	}
-	if field.Kind() != reflect.Bool {
-		return errors.New("field is not a bool")
-	}
-	field.SetBool(value)
-	return nil
-}
-
-// SetOptionInt sets an int option
-func (b *ExperimentBuilder) SetOptionInt(key string, value int64) error {
-	field, err := fieldbyname(b.config, key)
-	if err != nil {
-		return err
-	}
-	if field.Kind() != reflect.Int64 {
-		return errors.New("field is not an int64")
-	}
-	field.SetInt(value)
-	return nil
-}
-
-// SetOptionString sets a string option
-func (b *ExperimentBuilder) SetOptionString(key, value string) error {
-	field, err := fieldbyname(b.config, key)
-	if err != nil {
-		return err
-	}
-	if field.Kind() != reflect.String {
-		return errors.New("field is not a string")
-	}
-	field.SetString(value)
-	return nil
-}
-
-// SetCallbacks sets the interactive callbacks
-func (b *ExperimentBuilder) SetCallbacks(callbacks model.ExperimentCallbacks) {
-	b.callbacks = callbacks
-}
-
-func fieldbyname(v interface{}, key string) (reflect.Value, error) {
-	// See https://stackoverflow.com/a/6396678/4354461
-	ptrinfo := reflect.ValueOf(v)
-	if ptrinfo.Kind() != reflect.Ptr {
-		return reflect.Value{}, errors.New("value is not a pointer")
-	}
-	structinfo := ptrinfo.Elem()
-	if structinfo.Kind() != reflect.Struct {
-		return reflect.Value{}, errors.New("value is not a pointer to struct")
-	}
-	field := structinfo.FieldByName(key)
-	if !field.IsValid() || !field.CanSet() {
-		return reflect.Value{}, errors.New("no such field")
-	}
-	return field, nil
-}
-
-// NewExperiment creates the experiment
-func (b *ExperimentBuilder) NewExperiment() *Experiment {
-	experiment := b.build(b.config)
-	experiment.callbacks = b.callbacks
-	return experiment
-}
-
-// canonicalizeExperimentName allows code to provide experiment names
-// in a more flexible way, where we have aliases.
-func canonicalizeExperimentName(name string) string {
-	switch name = strcase.ToSnake(name); name {
-	case "ndt_7":
-		name = "ndt" // since 2020-03-18, we use ndt7 to implement ndt by default
-	default:
-	}
-	return name
-}
-
-func newExperimentBuilder(session *Session, name string) (*ExperimentBuilder, error) {
-	factory, _ := experimentsByName[canonicalizeExperimentName(name)]
-	if factory == nil {
-		return nil, fmt.Errorf("no such experiment: %s", name)
-	}
-	builder := factory(session)
-	builder.callbacks = model.NewPrinterCallbacks(session.Logger())
-	return builder, nil
 }
 
 // Experiment is an experiment instance.
@@ -437,244 +265,4 @@ func (e *Experiment) saveMeasurement(
 		return err
 	}
 	return filep.Close()
-}
-
-var experimentsByName = map[string]func(*Session) *ExperimentBuilder{
-	"dash": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, dash.NewExperimentMeasurer(
-					*config.(*dash.Config),
-				))
-			},
-			config:        &dash.Config{},
-			interruptible: true,
-			inputPolicy:   InputNone,
-		}
-	},
-
-	"example": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, example.NewExperimentMeasurer(
-					*config.(*example.Config), "example",
-				))
-			},
-			config: &example.Config{
-				Message:   "Good day from the example experiment!",
-				SleepTime: int64(time.Second),
-			},
-			interruptible: true,
-			inputPolicy:   InputNone,
-		}
-	},
-
-	"example_with_input": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, example.NewExperimentMeasurer(
-					*config.(*example.Config), "example_with_input",
-				))
-			},
-			config: &example.Config{
-				Message:   "Good day from the example with input experiment!",
-				SleepTime: int64(time.Second),
-			},
-			interruptible: true,
-			inputPolicy:   InputRequired,
-		}
-	},
-
-	// TODO(bassosimone): when we can set experiment options using the JSON
-	// we need to get rid of all these multiple experiments.
-	//
-	// See https://github.com/ooni/probe-engine/issues/413
-	"example_with_input_non_interruptible": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, example.NewExperimentMeasurer(
-					*config.(*example.Config), "example_with_input_non_interruptible",
-				))
-			},
-			config: &example.Config{
-				Message:   "Good day from the example with input experiment!",
-				SleepTime: int64(time.Second),
-			},
-			interruptible: false,
-			inputPolicy:   InputRequired,
-		}
-	},
-
-	"example_with_failure": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, example.NewExperimentMeasurer(
-					*config.(*example.Config), "example_with_failure",
-				))
-			},
-			config: &example.Config{
-				Message:     "Good day from the example with failure experiment!",
-				ReturnError: true,
-				SleepTime:   int64(time.Second),
-			},
-			interruptible: true,
-			inputPolicy:   InputNone,
-		}
-	},
-
-	"facebook_messenger": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, fbmessenger.NewExperimentMeasurer(
-					*config.(*fbmessenger.Config),
-				))
-			},
-			config:      &fbmessenger.Config{},
-			inputPolicy: InputNone,
-		}
-	},
-
-	"http_header_field_manipulation": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, hhfm.NewExperimentMeasurer(
-					*config.(*hhfm.Config),
-				))
-			},
-			config:      &hhfm.Config{},
-			inputPolicy: InputNone,
-		}
-	},
-
-	"http_invalid_request_line": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, hirl.NewExperimentMeasurer(
-					*config.(*hirl.Config),
-				))
-			},
-			config:      &hirl.Config{},
-			inputPolicy: InputNone,
-		}
-	},
-
-	"ndt": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, ndt7.NewExperimentMeasurer(
-					*config.(*ndt7.Config),
-				))
-			},
-			config:        &ndt7.Config{},
-			interruptible: true,
-			inputPolicy:   InputNone,
-		}
-	},
-
-	"psiphon": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, psiphon.NewExperimentMeasurer(
-					*config.(*psiphon.Config),
-				))
-			},
-			config:      &psiphon.Config{},
-			inputPolicy: InputOptional,
-		}
-	},
-
-	"sni_blocking": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, sniblocking.NewExperimentMeasurer(
-					*config.(*sniblocking.Config),
-				))
-			},
-			config: &sniblocking.Config{
-				ControlSNI: "example.com",
-			},
-			inputPolicy: InputRequired,
-		}
-	},
-
-	"stun_reachability": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, stunreachability.NewExperimentMeasurer(
-					*config.(*stunreachability.Config),
-				))
-			},
-			config:      &stunreachability.Config{},
-			inputPolicy: InputOptional,
-		}
-	},
-
-	"telegram": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, telegram.NewExperimentMeasurer(
-					*config.(*telegram.Config),
-				))
-			},
-			config:      &telegram.Config{},
-			inputPolicy: InputNone,
-		}
-	},
-
-	"tor": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, tor.NewExperimentMeasurer(
-					*config.(*tor.Config),
-				))
-			},
-			config:      &tor.Config{},
-			inputPolicy: InputNone,
-		}
-	},
-
-	"urlgetter": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, urlgetter.NewExperimentMeasurer(
-					*config.(*urlgetter.Config),
-				))
-			},
-			config:      &urlgetter.Config{},
-			inputPolicy: InputRequired,
-		}
-	},
-
-	"web_connectivity": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, webconnectivity.NewExperimentMeasurer(
-					*config.(*webconnectivity.Config),
-				))
-			},
-			config:      &webconnectivity.Config{},
-			inputPolicy: InputRequired,
-		}
-	},
-
-	"whatsapp": func(session *Session) *ExperimentBuilder {
-		return &ExperimentBuilder{
-			build: func(config interface{}) *Experiment {
-				return NewExperiment(session, whatsapp.NewExperimentMeasurer(
-					*config.(*whatsapp.Config),
-				))
-			},
-			config:      &whatsapp.Config{},
-			inputPolicy: InputNone,
-		}
-	},
-}
-
-// AllExperiments returns the name of all experiments
-func AllExperiments() []string {
-	var names []string
-	for key := range experimentsByName {
-		names = append(names, key)
-	}
-	return names
 }
